+++ Simulation/Pathfinding/ActionPathAux.hpp
@@ -2,7 +2,7 @@
 #define ACTION_PATH_AUX_HPP
 
 #include "Action.hpp"
-#include "_path.hpp"
+#include "Path.hpp"
 #include <memory>
 #include "Debugging.hpp"
 #include "GLOBALS.hpp"
