+++ Simulation/Pathfinding/Conflict.hpp
