+++ Simulation/Pathfinding/Edge.cpp
