+++ Simulation/Pathfinding/Edge.hpp
