+++ Simulation/Pathfinding/LowLevelCBS.hpp
@@ -5,7 +5,7 @@ class LowLevelCBS;
 
 #include <vector>
 #include <memory>
-#include "_path.hpp"
+#include "Path.hpp"
 #include "AgentInfo.hpp"
 #include "Action.hpp"
 #include "Debugging.hpp"
