+++ Simulation/Pathfinding/MapStructureGraph.hpp
