+++ Simulation/Pathfinding/Tests/CMakeLists.txt
@@ -2,8 +2,6 @@ include_directories(.)
 include_directories(./Unit/)
 include_directories(./Integration/)
 include_directories(..)
-include_directories(../ConstraintTree/)
-include_directories(../Interfaces/)
 include_directories(../../models/map/)
 include_directories(../../controllers/)
 
@@ -12,44 +10,36 @@ set(ControllersDir ../../controllers)
 set(MapDir ../../models/map)
 # We must be able to compile with these files alone (so map_structure wont work without argos files, which makes a lot of our files untestable atm)
 set(PathfindingFiles
-    ${PathfindingDir}/ConstraintTree/Conflict.hpp
-    ${PathfindingDir}/ConstraintTree/Conflict.cpp
-    ${PathfindingDir}/ConstraintTree/ConstraintTree.hpp
-    ${PathfindingDir}/ConstraintTree/ConstraintTree.cpp
+    ${PathfindingDir}/Conflict.hpp
+    ${PathfindingDir}/Conflict.cpp
+    ${PathfindingDir}/ConstraintTree.hpp
+    ${PathfindingDir}/ConstraintTree.cpp
     ${PathfindingDir}/Action.hpp
     ${PathfindingDir}/Action.cpp
     ${PathfindingDir}/ActionPathAux.hpp
     ${PathfindingDir}/ActionPathAux.cpp
     ${PathfindingDir}/BackwardsActionPathAux.hpp
     ${PathfindingDir}/BackwardsActionPathAux.cpp
-    # ${PathfindingDir}/ExperimentData.hpp
-    # ${PathfindingDir}/ExperimentData.cpp
-    # ${PathfindingDir}/Agent.hpp
-    # ${PathfindingDir}/Agent.cpp
     ${PathfindingDir}/AgentInfo.hpp
     ${PathfindingDir}/AgentInfo.cpp
-    ${PathfindingDir}/ConstraintTree/Constraint.hpp
-    ${PathfindingDir}/ConstraintTree/Constraint.cpp
+    ${PathfindingDir}/Constraint.hpp
+    ${PathfindingDir}/Constraint.cpp
     ${PathfindingDir}/HighLevelCBS.hpp
     ${PathfindingDir}/HighLevelCBS.cpp
     ${PathfindingDir}/LowLevelCBS.hpp
     ${PathfindingDir}/LowLevelCBS.cpp
-    ${PathfindingDir}/_path.hpp
-    ${PathfindingDir}/_path.cpp
+    ${PathfindingDir}/Path.hpp
+    ${PathfindingDir}/Path.cpp
     ${PathfindingDir}/Solution.hpp
     ${PathfindingDir}/Solution.cpp
     ${PathfindingDir}/Location.hpp
     ${PathfindingDir}/Location.cpp
-
-    # ${ControllersDir}/TestController.hpp
-    # ${ControllersDir}/TestController.cpp
-
-    ${PathfindingDir}/Interfaces/Edge.hpp
-    ${PathfindingDir}/Interfaces/Edge.cpp
-    ${PathfindingDir}/Interfaces/Graph.hpp
-    ${PathfindingDir}/Interfaces/Graph.cpp
-    ${PathfindingDir}/Interfaces/Vertex.hpp
-    ${PathfindingDir}/Interfaces/Vertex.cpp
+    ${PathfindingDir}/Edge.hpp
+    ${PathfindingDir}/Edge.cpp
+    ${PathfindingDir}/Graph.hpp
+    ${PathfindingDir}/Graph.cpp
+    ${PathfindingDir}/Vertex.hpp
+    ${PathfindingDir}/Vertex.cpp
 
     ${PathfindingDir}/Debugging.hpp
     ${PathfindingDir}/Logger.hpp
