+++ Simulation/Pathfinding/Vertex.cpp
