+++ Simulation/Pathfinding/Vertex.hpp
